Fix: data type of the weight float to double

+++ Connection.cpp
@@ -1,8 +1,8 @@
 #include "Connection.h"
 
-Connection::Connection(int id, int origin, int destination, float weight)
+Connection::Connection(int id, int origin, int destination, double weight)
     : id(id), origin(origin), destination(destination), weight(weight) {}
 
-void Connection::actualizarPeso(float nuevoPeso) {
+void Connection::actualizarPeso(double nuevoPeso) {
     weight = nuevoPeso;
 }
