+++ Node.cpp
@@ -1,6 +1,6 @@
 #include "Node.h"
 
-void Node::actualizarRelevancia(float nuevaRelevancia) {
+void Node::actualizarRelevancia(double nuevaRelevancia) {
     relevanciaPrev = relevanciaActual;
     relevanciaActual = nuevaRelevancia;
 }
